test(store): move setting store tests to tests-ui/tests/store/settingStore.test.ts
